Fix for issue 2870: status and dsreplication status do not show replication monitoring The changes consist basically in using the new attributes "approx-older-change-not-synchronized-millis" and "missing-changes" in status and replication status command-lines.  In fact a millisecond version of approx-older-change-not-synchronized-millis has been added in order the remote command-line to be able to provide a localized version of the date (instead of directly presenting the date that was written by the server).
The code has been updated to take into consideration the fact that there is information for several servers; so the replication domain id is used to discriminate the information of the different servers.

+++ opendj-sdk/opends/src/ads/org/opends/admin/ads/ReplicaDescriptor.java
@@ -22,7 +22,7 @@
  * CDDL HEADER END
  *
  *
- *      Portions Copyright 2007 Sun Microsystems, Inc.
+ *      Portions Copyright 2007-2008 Sun Microsystems, Inc.
  */
 
 package org.opends.admin.ads;
@@ -42,7 +42,7 @@ public class ReplicaDescriptor
   private Set<String> replicationServers = new HashSet<String>();
   private int replicationId = -1;
   private int missingChanges = -1;
-  private int ageOfOldestMissingChange = -1;
+  private long ageOfOldestMissingChange = -1;
 
   /**
    * Returns the number of entries contained in the replica.
@@ -167,7 +167,7 @@ public class ReplicaDescriptor
    * Returns the age of the oldest missing change.
    * @return the age of the oldest missing change.
    */
-  public int getAgeOfOldestMissingChange()
+  public long getAgeOfOldestMissingChange()
   {
     return ageOfOldestMissingChange;
   }
@@ -176,7 +176,7 @@ public class ReplicaDescriptor
    * Sets the age of the oldest missing change.
    * @param ageOfOldestMissingChange the age of the oldest missing change.
    */
-  public void setAgeOfOldestMissingChange(int ageOfOldestMissingChange)
+  public void setAgeOfOldestMissingChange(long ageOfOldestMissingChange)
   {
     this.ageOfOldestMissingChange = ageOfOldestMissingChange;
   }
